+++ backend/packages/gitlab/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @cat-factory/gitlab
 
+## 0.4.22
+
+### Patch Changes
+
+- Updated dependencies [77c6842]
+  - @cat-factory/contracts@0.71.0
+  - @cat-factory/kernel@0.63.3
+
 ## 0.4.21
 
 ### Patch Changes